Addressed PR comments and restructured/refactored code to reduce code duplication across e2e tests
