+++ test/cypress-e2e/README.md
@@ -43,7 +43,11 @@ cypress.env.json
 
 ## Writing Tests
 
-Test files are located in the `cypress/e2e/` directory. Each test file should follow the `.cy.js` naming convention. Example test:
+Test files are located in the `cypress/e2e/` directory. Each test file should follow the `.cy.js` naming convention. 
+
+Currently, we have organized tests such that large e2e flows are broken down into respective page objects and added under "support > pages". This allows us to reuse these page components to minimize code duplication. 
+
+These components can then be called in the e2e tests to replicate a full user flow (refer to group-application-full-e2e.cy.js for an example).
 
 ## Run tests:
 
@@ -55,14 +59,17 @@ or for headless mode:
       npx cypress run
 
 ### Group Application:
+To run the full E2E automation, navigate to "e2e > 2-group-application-full-e2e".
 
-To run the full E2E automation, run the "full-e2e.cy.js" file when opening cypress.
 There are 3 areas requiring future contribution for added automation functionality:
 1) Adding Multiple Facilities to a single group application
 2) Adding Multiple Closures to a facility
 3) Adding a Partial Closure affecting specific Care Categories in a facility
 
-These functions are more challenging to implement due to the structure of their respective pages in the Portal. Each area has been marked with "TODO" in terms of where it would likely need to be added in all scripts under "2-group-application".
+These functions are more challenging to implement due to the structure of their respective pages in the Portal. Each area has been marked with "TODO" in terms of where they would likely need to be added in their respective scripts (located under pages > 2-group-application-pages).
+
+If you choose to work on them, please update the comment with your github username.
+
 
 ## Contributing
