+++ test/cypress-e2e/cypress/e2e/1-portal-login/portal-login-invalid-credentials-e2e.cy.js
@@ -1,4 +1,4 @@
-import { loginPage } from "../../support/pages/LoginPage";
+import { loginPage } from "../../support/pages/login-page/LoginPage";
 
 describe("Login Page Test", () => {
   it("should login successfully with valid credentials", () => {
